dnt/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.6
django-widget-tweaks==1.4.5
idna==2.8
defusedxml==0.6.0
olefile==0.46
Pillow==6.0.0
pytesseract==0.3.0
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
setuptools==41.0.1
wheel==0.33.1
argon2-cffi==19.1.0

